refactor: extract types and sort utils to pass file size check
- analysis.ts (428→305): types moved to analysis-types.ts
- TranscriptList.tsx (630→374): sort utils to transcriptListSort.ts,
  header components to SortHeaderButton.tsx
